set specific available historical ranges for certain layers
