+++ config/default/common/config/wv.json/layers/goes/GOES-East_ABI_Air_Mass.json
@@ -16,7 +16,9 @@
         "day",
         "night"
       ],
-      "availableWindow": 30
+      "availability": {
+        "rollingWindow": 30
+      }
     }
   }
 }
