+++ config/default/common/config/wv.json/layers/goes/GOES-East_ABI_Band13_Clean_Infrared.json
@@ -19,7 +19,16 @@
         "day",
         "night"
       ],
-      "availableWindow": 30
+      "availability": {
+        "rollingWindow": 30,
+        "historicalRanges": [
+          {
+            "startDate": "2019-09-01T10:00:00Z",
+            "endDate": "2019-09-05T18:00:00Z",
+            "dateInterval": "10"
+          }
+        ]
+      }
     }
   }
 }
