+++ config/default/common/config/wv.json/layers/goes/GOES-West_ABI_Band13_Clean_Infrared.json
@@ -19,7 +19,9 @@
         "day",
         "night"
       ],
-      "availableWindow": 30
+      "availability": {
+        "rollingWindow": 30
+      }
     }
   }
 }
