+++ config/default/common/config/wv.json/layers/goes/GOES-West_ABI_Band2_Red_Visible_1km.json
@@ -15,7 +15,9 @@
       "daynight": [
         "day"
       ],
-      "availableWindow": 30
+      "availability": {
+        "rollingWindow": 30
+      }
     }
   }
 }
