+++ config/default/common/config/wv.json/layers/himawari/Himawari_AHI_Band13_Clean_Infrared.json
@@ -19,7 +19,16 @@
         "day",
         "night"
       ],
-      "availableWindow": 30
+      "availability": {
+        "rollingWindow": 30,
+        "historicalRanges": [
+          {
+            "startDate": "2020-01-03",
+            "endDate": "2020-01-04",
+            "dateInterval": "10"
+          }
+        ]
+      }
     }
   }
 }
