+++ config/default/common/config/wv.json/layers/himawari/Himawari_AHI_Band3_Red_Visible_1km.json
@@ -15,7 +15,16 @@
       "daynight": [
         "day"
       ],
-      "availableWindow": 30
+      "availability": {
+        "rollingWindow": 30,
+        "historicalRanges": [
+          {
+            "startDate": "2020-01-03",
+            "endDate": "2020-01-04",
+            "dateInterval": "10"
+          }
+        ]
+      }
     }
   }
 }
